+++ web/package.json
@@ -91,6 +91,7 @@
     "gatsby-plugin-page-creator": "^4.12.1",
     "gatsby-plugin-react-helmet": "^5.12.1",
     "gatsby-plugin-react-leaflet": "^3.0.3",
+    "gatsby-plugin-react-svg": "^3.1.0",
     "gatsby-plugin-resolve-src": "^2.1.0",
     "gatsby-plugin-sharp": "^4.12.1",
     "gatsby-source-filesystem": "^4.12.1",
@@ -121,7 +122,7 @@
     "remark-parse": "^10.0.1",
     "remark-rehype": "^10.1.0",
     "scroll-to-element": "^2.0.3",
-    "sharp": ">=0.30",
+    "sharp": "^0.30.4",
     "start": "^5.1.0",
     "swiper": "^8.1.3",
     "unified": "^10.1.2",
