--- web/src/contents/AboutUsContent/data/index.js
@@ -1,239 +1,0 @@
-// assets
-import Person1 from 'assets/images/about/allen.jpeg';
-import Person2 from 'assets/images/about/brett.jpg';
-import Person3 from 'assets/images/about/michael.jpg';
-import Person4 from 'assets/images/about/nathan.jpeg';
-import Person5 from 'assets/images/about/tasha.jpeg';
-
-export const team = [
-  {
-    title: 'UI/UX Designer',
-    authorPhoto: {
-      src: 'https://assets.maccarianagency.com/the-front/photos/people/kate-segelson.jpg',
-      srcSet:
-        'https://assets.maccarianagency.com/the-front/photos/people/kate-segelson@2x.jpg 2x',
-    },
-    authorName: 'Kate Segelson',
-  },
-  {
-    title: 'Web Backend Developer',
-    authorPhoto: {
-      src: 'https://assets.maccarianagency.com/the-front/photos/people/alex-johnson.jpg',
-      srcSet:
-        'https://assets.maccarianagency.com/the-front/photos/people/alex-johnson@2x.jpg 2x',
-    },
-    authorName: 'Alex Johnson',
-  },
-  {
-    title: 'Web Frontend Developer',
-    authorPhoto: {
-      src: 'https://assets.maccarianagency.com/the-front/photos/people/valeria-kogan.jpg',
-      srcSet:
-        'https://assets.maccarianagency.com/the-front/photos/people/valeria-kogan@2x.jpg 2x',
-    },
-    authorName: 'Valeria Kogan',
-  },
-  {
-    title: 'CEO / Co-Founder',
-    authorPhoto: {
-      src: 'https://assets.maccarianagency.com/the-front/photos/people/akachi-luccini.jpg',
-      srcSet:
-        'https://assets.maccarianagency.com/the-front/photos/people/akachi-luccini@2x.jpg 2x',
-    },
-    authorName: 'Akachi Luccini',
-  },
-  {
-    title: 'CTO / Co-Founder',
-    authorPhoto: {
-      src: 'https://assets.maccarianagency.com/the-front/photos/people/jack-smith.jpg',
-      srcSet:
-        'https://assets.maccarianagency.com/the-front/photos/people/jack-smith@2x.jpg 2x',
-    },
-    authorName: 'Jack Smith',
-  },
-  {
-    title: 'Consultant',
-    authorPhoto: {
-      src: 'https://assets.maccarianagency.com/the-front/photos/people/veronica-adams.jpg',
-      srcSet:
-        'https://assets.maccarianagency.com/the-front/photos/people/veronica-adams@2x.jpg 2x',
-    },
-    authorName: 'Veronica Adams',
-  },
-];
-
-export const companies = [
-  {
-    name: 'Airbnb',
-    logo: 'https://assets.maccarianagency.com/the-front/logos/airbnb.svg',
-  },
-  {
-    name: 'Coinbase',
-    logo: 'https://assets.maccarianagency.com/the-front/logos/coinbase.svg',
-  },
-  {
-    name: 'Dribbble',
-    logo: 'https://assets.maccarianagency.com/the-front/logos/dribbble.svg',
-  },
-];
-
-export const mapData = [
-  {
-    location: {
-      y: 45.453211,
-      x: 9.248383,
-      address: 'Via A.G. Alaimo 147, 55027',
-    },
-  },
-  {
-    location: {
-      y: 45.419211,
-      x: 9.021383,
-      address: 'Via Rocca de Baldi 95, 440368',
-    },
-  },
-  {
-    location: {
-      y: 45.473211,
-      x: 9.298383,
-      address: 'Via Firenze 134, 45588',
-    },
-  },
-  {
-    location: {
-      y: 45.461211,
-      x: 9.000383,
-      address: 'Via Cavour 29, 201558',
-    },
-  },
-  {
-    location: {
-      y: 45.413211,
-      x: 9.398383,
-      address: 'Via Bologna 33, 220156',
-    },
-  },
-  {
-    location: {
-      y: 45.569211,
-      x: 9.128383,
-      address: 'Vicolo Tre Marchetti 127, 350125',
-    },
-  },
-  {
-    location: {
-      y: 45.483211,
-      x: 9.148383,
-      address: 'Via Lombardi 146, 45830',
-    },
-  },
-  {
-    location: {
-      y: 45.313211,
-      x: 9.012383,
-      address: 'Via Guantai Nuovi 99, 56989',
-    },
-  },
-];
-
-export const gallery = [
-  {
-    image: {
-      src: 'https://assets.maccarianagency.com/the-front/photos/expo-gallery/gallery1.jpg',
-      srcSet:
-        'https://assets.maccarianagency.com/the-front/photos/expo-gallery/gallery1@2x.jpg 2x',
-    },
-    title: 'gallery',
-    cols: 1,
-  },
-  {
-    image: {
-      src: 'https://assets.maccarianagency.com/the-front/photos/expo-gallery/gallery3.jpg',
-      srcSet:
-        'https://assets.maccarianagency.com/the-front/photos/expo-gallery/gallery3@2x.jpg 2x',
-    },
-    title: 'gallery',
-    cols: 3,
-  },
-  {
-    image: {
-      src: 'https://assets.maccarianagency.com/the-front/photos/expo-gallery/gallery2.jpg',
-      srcSet:
-        'https://assets.maccarianagency.com/the-front/photos/expo-gallery/gallery2@2x.jpg 2x',
-    },
-    title: 'gallery',
-    cols: 3,
-  },
-  {
-    image: {
-      src: 'https://assets.maccarianagency.com/the-front/photos/coworking/place3.jpg',
-      srcSet:
-        'https://assets.maccarianagency.com/the-front/photos/coworking/place3@2x.jpg 2x',
-    },
-    title: 'gallery',
-    cols: 1,
-  },
-];
-
-export const paragraphs = {
-  our_story:
-    "Angel Scale, LLC, and Startup Resource Center (SRC) were created to support those who are relatively new to starting, owning, and running a successful business. Founded in March of 2021, Angel Scale began by helping five new businesses hit the ground running: Grey and Bash Teas -- a loose-leaf tea company, Honey Sweet Collective -- a company that manufactures and sells fun coffee mugs, Insomnia PJ -- a high-end line of women's loungewear, Hot Fit NYC -- a Mixed Martial Arts fitness facility in NYC, and Medical Doorway -- an International Education firm that helps high school graduates enroll in Medical schools in six countries throughout Europe and Asia.  Each of these companies is now thriving, thanks in part to the resources, insight, and support offered by Angel Scale’s business experts.",
-  who_are_we:
-    'Angel Scale is committed to helping startup and pre-startup entrepreneurs navigate the path from ideation to activation; from startup, to profitability, and beyond. Startup Resource Center was created as part of Angel Scale’s arsenal of resources to expedite entrepreneurs to success, offering new entrepreneurs the information, data, and actual resources needed to take their idea to market, all through an easy-to-access digital site. SRC further recognizes that great ideas are only a starting point, and thus aims to connect entrepreneurs with people and companies that can provide the tools, software, and expertise needed to pave the way to success.',
-  our_process:
-    'Angel Scale and SRC are run by a small team of experts with first-hand experience in entrepreneurial pursuits. As business owners, we understand the excitement, concerns, and challenges new startups face, and we want to help. Angel Scale is committed to taking young entrepreneurs through the process of starting and growing a business.  We offer different levels of support, from valuable content and product/service reviews, to one-on-one consulting relationships. Each of us have been there, facing the obstacles that come with starting and developing new businesses, so we know what to expect and can serve as navigators for those entering what could otherwise be a bumpy road.  We fully believe in the path from living room to CEO suite, and that, with proper guidance, any strong idea can become a successful business entity.',
-};
-
-export const newTeam = [
-  {
-    title: 'Partner',
-    authorPhoto: {
-      src: Person2,
-      srcSet: Person2,
-    },
-    authorName: 'Brett Ozar',
-    bioOne: `Brett Ozar started her career in the fashion industry, operating sales for familiar brands like Union Bay. She became Marc Ecko’s right hand woman in licensing, launching, managing, and marketing Ecko's product division. Over her time at Ecko, she grew the licensed product division to 12 product categories. From there, Ozar became a consultant for startup businesses and companies looking to rebrand, using her expansive knowledge and experience in marketing, sales, and product variety to apply universal winning concepts to a broad range of market categories, helping brands build new audiences through successful launches. One of her most recent passion projects is a partnership with LiveNation on Fleaesta: a curated fleamarket/fiesta that combines Ozar's love for vintage goods with a platform where artists, makers, collectors, and entrepreneurs can connect.`,
-    bioTwo: `Among friends and colleagues, Ozar is known as the go-to advice person. She takes pride in building ideas up to the next level and connecting eager entrepreneurs with the right contacts. Her passion and expertise lie in building networks and developing concepts, which makes her an excellent addition to the Angel-Scale family, providing clear insight to budding entrepreneurs who aren't sure where to start.`,
-  },
-  {
-    title: 'Partner',
-    authorPhoto: {
-      src: Person3,
-      srcSet: Person3,
-    },
-    authorName: 'Michael Platt',
-    bioOne: `Michael Platt is a serial entrepreneur, and earned the distinction of Ernst and Young's "Entrepreneur of the Year". Having built and sold successful businesses in the 80's and early 90's including a fundraising restaurant discount card, a professional wrestling organization, a fine photography online art store, a chain of schools that trained people to become personal fitness trainers and massage therapists, and an advertising agency, Platt embarked on what was to be his final entrepreneurial venture, an education marketing firm called PlattForm Advertising (now Keypath Education).  `,
-    bioTwo: `From a humble beginning in the basement of his home, Platt built the marketing firm to  over 250 million in revenue with over 400 employees and offices in the US, Canada, the UK and Australia, before taking it public on the ASX, the Australian stock market.  Ultimately, this final venture was far from final as Platt found himself joining boards of directors and consulting friends and family, and decided retirement was not for him.  He pulled together a team of people he knew, trusted and whose opinions and successes he valued, and formed Angel Scale, LLC.`,
-  },
-  {
-    title: 'Partner',
-    authorPhoto: {
-      src: Person5,
-      srcSet: Person5,
-    },
-    authorName: 'Tasha Cerny',
-    posTop: true,
-    bioOne: `Tasha Cerny has been developing and managing a variety of content resources for more than a decade. With writing as a lifelong passion, Cerny has pursued content creation in every form: from marketing and publicity to copywriting and social media management. While also establishing a career in the entertainment industry as a television writer, Cerny has built several entrepreneurial side businesses and projects, including a successful Etsy resale shop and producing a livestream interview show, Pendemictv, that showcased the wisdom of over sixty well-known, successful screenwriters. `,
-    bioTwo: `Over the years Cerny has built a reputation as a versatile writer and team builder. Cerny currently works with the executive producers of ABC’s hit drama, The Good Doctor, in addition to utilizing her expertise as part of the Angel-Scale team. She looks forward to sharing everything she’s learned about starting businesses and passion projects with Angel-Scale.`,
-  },
-  {
-    title: 'Partner',
-    authorPhoto: {
-      src: Person4,
-      srcSet: Person4,
-    },
-    authorName: 'Nathan Cerny',
-    bioOne: `Nathan Cerny got his start in the technology space at a young age, creating his own business building and fixing computers while still in high school. After graduating from Kansas State University with a B.S. in Electrical Engineering and Digital Electronics, Cerny spent the next fifteen years in the software and information technology space, honing his skills in digital security and quickly climbing the ladder of corporate leadership. Unsatisfied working within the limits of one company’s digital space, Cerny branched off to build his own consulting firm, where he focuses on bringing his insight and experience to the technological and digital security needs of any business.`,
-    bioTwo: `In addition to running his own business, Cerny played a vital part in the creation of Angel-Scale, building the very site you’re reading today and ensuring that every aspect of our core values and purpose are met through our digital platform. `,
-  },
-  {
-    title: 'Partner',
-    authorPhoto: {
-      src: Person1,
-      srcSet: Person1,
-    },
-    authorName: 'Allen Harkleroad',
-    bioOne: `For more than 15 years, Allen Harkleroad has been building teams within the marketing space to prepare businesses for the future. Harkleroad began his career coordinating operations for PlattForm Advertising. He then worked his way up to lead the department, where he was integral in shifting operations from simple project coordination to full project management. From there, he transitioned focus to search engine optimization, where he built a team and expanded it to include content development, social media management, and user experience optimization. Under Harkleroad's leadership, the team grew into one of the largest and most profitable departments at Thruline Marketing. Harkleroad then became vice president of internet marketing at The frank Agency where he oversaw all internet marketing before bringing his expertise in management, tool integration, and employee training to Angel Scale, LLC as COO.`,
-    bioTwo: `From a degree in archaeology, to teaching English in Japan and working as a surf instructor in Hawaii, Harkleroad has had an unconventional journey to his rise in the marketing world. His broad and unique insight into organizations and people has made him an asset to businesses in every market. He is excited to bring those insights to Angel Scale, where he can share his experiences and expertise with others and work with a new team of talented business experts.`,
-  },
-];
